Merge branch 'fix-issue-400' into 'main'

Closes #400

For details see https://gitlab.com/hpctoolkit/hpcviewer/-/merge_requests/381

Approved-by: John Mellor-Crummey <johnmc@rice.edu>

+++ edu.rice.cs.hpctraceviewer.ui/src/edu/rice/cs/hpctraceviewer/ui/main/SpaceTimeDetailCanvas.java
@@ -682,7 +682,9 @@ public class SpaceTimeDetailCanvas extends AbstractTimeCanvas
     	final Position position = stData.getTraceDisplayAttribute().getPosition();    			
 		final TimeUnit dbTimeUnit = stData.getTimeUnit();
 		final TimeUnit displayTimeUnit = attributes.getTimeUnit();
-		final float selectedTime = displayTimeUnit.convert(position.time, dbTimeUnit) * attributes.getTimeUnitMultiplier();
+		
+		// fix issue #400: convert to double to ensure we don't lose the precision 
+		final double selectedTime = displayTimeUnit.convert(position.time, dbTimeUnit) * (double)attributes.getTimeUnitMultiplier();
 
         final var traceData = stData.getBaseData();
